+++ components/faphub/dao/api/src/main/java/com/flipperdevices/faphub/dao/api/FapVersionApi.kt
@@ -1,7 +1,7 @@
 package com.flipperdevices.faphub.dao.api
 
-import com.flipperdevices.faphub.dao.api.model.FapItemVersion
+import com.flipperdevices.core.data.SemVer
 
 interface FapVersionApi {
-    suspend fun getVersions(versions: List<String>): List<FapItemVersion>
+    suspend fun getVersionsMap(versions: List<String>): Map<String, SemVer>
 }