+++ components/faphub/installation/button/impl/src/main/java/com/flipperdevices/faphub/installation/button/impl/api/FapInstallationUIApiImpl.kt
@@ -79,6 +79,17 @@ class FapInstallationUIApiImpl @Inject constructor() : FapInstallationUIApi {
                 fapButtonSize = fapButtonSize,
                 modifier = modifier
             )
+
+            FapState.ConnectFlipper -> ComposableFapInstallButton(
+                modifier = modifier,
+                fapButtonSize = fapButtonSize,
+                onClick = { }
+            )
+
+            FapState.FlipperOutdated -> ComposableFapInstalledButton(
+                modifier = modifier,
+                fapButtonSize = fapButtonSize
+            )
         }
     }
 }