+++ components/faphub/installation/button/impl/src/main/java/com/flipperdevices/faphub/installation/button/impl/viewmodel/FapStatusViewModel.kt
@@ -34,7 +34,9 @@ class FapStatusViewModel @VMInject constructor(
                 applicationAlias = fapButtonConfig.applicationAlias,
                 applicationUid = fapButtonConfig.applicationUid,
                 toVersion = fapButtonConfig.version,
-                categoryAlias = fapButtonConfig.categoryAlias
+                categoryAlias = fapButtonConfig.categoryAlias,
+                applicationName = fapButtonConfig.applicationName,
+                iconUrl = fapButtonConfig.iconUrl
             )
         )
     }
@@ -56,7 +58,9 @@ class FapStatusViewModel @VMInject constructor(
         queueApi.enqueue(
             FapActionRequest.Update(
                 from = from,
-                toVersion = fapButtonConfig.version
+                toVersion = fapButtonConfig.version,
+                iconUrl = fapButtonConfig.iconUrl,
+                applicationName = fapButtonConfig.applicationName
             )
         )
     }