+++ components/faphub/installation/manifest/api/src/main/java/com/flipperdevices/faphub/installation/manifest/api/FapManifestApi.kt
@@ -1,11 +1,12 @@
 package com.flipperdevices.faphub.installation.manifest.api
 
-import com.flipperdevices.faphub.installation.manifest.model.FapManifestItem
+import com.flipperdevices.faphub.installation.manifest.model.FapManifestEnrichedItem
+import com.flipperdevices.faphub.installation.manifest.model.FapManifestState
 import kotlinx.coroutines.flow.StateFlow
 
 interface FapManifestApi {
-    fun getManifestFlow(): StateFlow<List<FapManifestItem>?>
-    suspend fun add(pathToFap: String, fapManifestItem: FapManifestItem)
-    suspend fun remove(applicationId: String)
+    fun getManifestFlow(): StateFlow<FapManifestState>
+    suspend fun add(pathToFap: String, fapManifestItem: FapManifestEnrichedItem)
+    suspend fun remove(applicationUid: String)
     fun invalidateAsync()
 }