+++ components/faphub/installation/manifest/api/src/main/java/com/flipperdevices/faphub/installation/manifest/model/FapManifestItem.kt
@@ -1,8 +1,13 @@
 package com.flipperdevices.faphub.installation.manifest.model
 
+import com.flipperdevices.core.data.SemVer
+
 data class FapManifestItem(
     val applicationAlias: String,
     val uid: String,
-    val version: FapManifestVersion,
-    val path: String
+    val versionUid: String,
+    val path: String,
+    val fullName: String,
+    val iconBase64: String?,
+    val sdkApi: SemVer?
 )