+++ components/faphub/installation/manifest/api/src/main/java/com/flipperdevices/faphub/installation/manifest/model/FapManifestState.kt
@@ -1,0 +1,17 @@
+package com.flipperdevices.faphub.installation.manifest.model
+
+import kotlinx.collections.immutable.ImmutableList
+
+sealed class FapManifestState {
+    object Loading : FapManifestState()
+
+    object NotLoaded : FapManifestState()
+
+    data class LoadedOffline(
+        val items: ImmutableList<FapManifestItem>
+    ) : FapManifestState()
+
+    data class Loaded(
+        val items: ImmutableList<FapManifestEnrichedItem>
+    ) : FapManifestState()
+}