+++ components/faphub/installation/manifest/impl/src/main/java/com/flipperdevices/faphub/installation/manifest/impl/model/FapManifestInternalItem.kt
@@ -1,8 +1,0 @@
-package com.flipperdevices.faphub.installation.manifest.impl.model
-
-data class FapManifestInternalItem(
-    val applicationAlias: String,
-    val uid: String,
-    val versionUid: String,
-    val path: String
-)