+++ components/faphub/installation/manifest/impl/src/main/java/com/flipperdevices/faphub/installation/manifest/impl/utils/FapManifestsLoader.kt
@@ -9,9 +9,9 @@ import com.flipperdevices.core.ktx.jre.flatten
 import com.flipperdevices.core.ktx.jre.pmap
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
-import com.flipperdevices.faphub.installation.manifest.impl.model.FapManifestInternalItem
 import com.flipperdevices.faphub.installation.manifest.impl.utils.FapManifestConstants.FAP_MANIFESTS_FOLDER_ON_FLIPPER
 import com.flipperdevices.faphub.installation.manifest.impl.utils.FapManifestConstants.FAP_MANIFEST_EXTENSION
+import com.flipperdevices.faphub.installation.manifest.model.FapManifestItem
 import com.flipperdevices.protobuf.main
 import com.flipperdevices.protobuf.storage.listRequest
 import com.flipperdevices.protobuf.storage.readRequest
@@ -26,7 +26,7 @@ class FapManifestsLoader @Inject constructor(
 ) : LogTagProvider {
     override val TAG = "FapManifestsLoader"
 
-    suspend fun load(): List<FapManifestInternalItem> {
+    suspend fun load(): List<FapManifestItem> {
         val serviceApi = flipperServiceProvider.getServiceApi()
         info { "Start load manifests" }
         var manifestNames = getManifestPaths(serviceApi.requestApi)